src/connectors/youtube_uploader.py
"""
YouTube Uploader Module
Handles uploading content to YouTube using YouTube Data API
"""

from googleapiclient.discovery import build
from google.oauth2.credentials import Credentials
from typing import Dict, Any, Optional
import logging

logger = logging.getLogger(__name__)


def upload_to_youtube(file_path: str, title: str, description: str, tags: list = None) -> Dict[str, Any]:
    """
    Upload video to YouTube using YouTube Data API
    
    Args:
        file_path (str): Path to the video file to upload
        title (str): Video title
        description (str): Video description
        tags (list): List of tags for the video
        
    Returns:
        Dict[str, Any]: Result of the upload operation
    """
    # Implementation will be added in Phase 3
    pass


def authenticate_youtube() -> Optional[object]:
    """
    Authenticate with YouTube Data API
    
    Returns:
        Optional[object]: YouTube service object if authentication successful
    """
    # Implementation will be added in Phase 3
    pass


def validate_youtube_credentials() -> bool:
    """
    Validate YouTube API credentials
    
    Returns:
        bool: True if credentials are valid, False otherwise
    """
    # Implementation will be added in Phase 3
    pass